main is green again: three things #3563 and my own #3613 left red
Three breaks, none of which any single PR's own gate could see.

waitingScanCap was exported to WaitingScanCap by #3563 and one call site
kept the old name, so the backend did not compile at all. A comment two
hundred lines up kept the old spelling too.

The same commit grew waiting.go to 563 lines, past the 500 cap. The three
functions it added answer a different question from the rest of the file
— narrowing the who-is-waiting read to ONE record, for the record page's
panel, rather than to the workspace — so they move to waitingrecord.go
rather than take a waiver.

And #3613, mine, put a module integration test where depguard forbids it:
a module may never import the composition layer, and that test drives the
compose harness. It moves beside the harness, which is where every other
module test needing a seeded workspace already lives. I did not see it
because the golangci lock was contended when I gated that branch and I
shipped on the scoped run — the scoped run does not evaluate depguard's
module list.

All three merged green on their own branches and only collide in the
tree, which is the case a per-PR gate cannot catch.
