+++ backend/internal/modules/activities/waiting.go
@@ -30,7 +30,6 @@ import (
 	"github.com/margince/margince/backend/internal/platform/auth"
 	"github.com/margince/margince/backend/internal/shared/kernel/ids"
 	"github.com/margince/margince/backend/internal/shared/kernel/principal"
-	"github.com/margince/margince/backend/internal/shared/ports/datasource"
 )
 
 // WaitingReply is one inbound message nobody has answered.
@@ -248,8 +247,8 @@ const waitingRepliesSQL = `
 	   AND a.archived_at IS NULL
 	   AND a.occurred_at <= $%[1]d
 	   AND %[2]s
-	   -- Entity narrowing goes HERE, before waitingScanCap's LIMIT below: a
-	   -- record's own wait can sit outside the oldest waitingScanCap threads
+	   -- Entity narrowing goes HERE, before WaitingScanCap's LIMIT below: a
+	   -- record's own wait can sit outside the oldest WaitingScanCap threads
 	   -- workspace-wide, and narrowing after the cap would report nothing
 	   -- waiting on the very record this asks about. "TRUE" for the
 	   -- workspace-wide Worklist read.
@@ -483,81 +482,3 @@ func (s *Store) WaitingReplies(ctx context.Context, asOf time.Time) ([]WaitingRe
 
 // waitingReplyEntityClause narrows the thread walk to one record, in the SAME
 // vocabulary linktarget.go and listActivitiesFilter's own entity_type/id
-// filter use — a record type added to linkColumn or the organization arm
-// reaches this walk too, rather than a second copy silently missing it.
-func waitingReplyEntityClause(entityType string, entityID ids.UUID, arg func(any) int) (string, error) {
-	if entityType == string(datasource.RecordOrganization) {
-		// An account's timeline is wider than its direct links (mail is filed
-		// against the person it was with), so this reuses the SAME three-arm
-		// walk the timeline list and the company view both read through —
-		// see OrgLinkedActivityExists.
-		return OrgLinkedActivityExists(arg(entityID)), nil
-	}
-	column := linkColumn(entityType)
-	if column == "" {
-		return "", &InvalidLinkTypeError{EntityType: entityType}
-	}
-	typePos := arg(entityType)
-	idPos := arg(entityID)
-	return sprintf("EXISTS (SELECT 1 FROM activity_link el WHERE el.activity_id = a.id AND el.entity_type = $%d AND el.%s = $%d)",
-		typePos, column, idPos), nil
-}
-
-// waitingReplyExistsClause builds the timeline list's `waiting_reply=true`
-// filter: the SAME thread walk WaitingReplies runs for the Worklist,
-// embedded as a subquery so the outer list's own entity/kind/cursor terms
-// compose with it rather than duplicating what "unanswered" means.
-//
-// The subquery is uncorrelated — it computes its own candidate set rather
-// than reading the outer FROM — so its own `a` alias shadowing the outer
-// query's is harmless.
-func waitingReplyExistsClause(ctx context.Context, arg func(any) int, asOf time.Time, entityType *string, entityID *ids.UUID) (string, error) {
-	instant := arg(asOf)
-	content, err := auth.ActivityContentClause(ctx, "a", arg)
-	if err != nil {
-		return "", err
-	}
-	linkVisible, err := auth.LinkTargetVisibleClause(ctx, "wl", arg)
-	if err != nil {
-		return "", err
-	}
-	if linkVisible == "" {
-		linkVisible = scopeUnbounded
-	}
-	entityClause := scopeUnbounded
-	if entityType != nil && entityID != nil {
-		entityClause, err = waitingReplyEntityClause(*entityType, *entityID, arg)
-		if err != nil {
-			return "", err
-		}
-	}
-	// The SAME reader, horizon and live-record rules WaitingReplies applies
-	// for the Worklist: a thread the Worklist would not name as waiting must
-	// not be named by a record page either, and a per-record set-aside must
-	// still be this reader's own.
-	reader := arg(readerOrNobody(ctx))
-	return "a.id IN (SELECT id FROM (" +
-		fmt.Sprintf(waitingRepliesSQL, instant, content, linkVisible, waitingScanCap,
-			waitingHorizonDays,
-			liveRecord(openDealPredicate, "d"),
-			liveRecord(workingLeadPredicate, "ld"),
-			liveRecord(openDealPredicate, "openDeal"),
-			liveRecord(openDealPredicate, "fd"),
-			reader,
-			entityClause) +
-		") waiting_thread)", nil
-}
-
-// appendWaitingReplyClause is listActivitiesFilter's `waiting_reply=true`
-// term, split out so that already-long function does not have to grow to
-// hold it. A no-op when the caller did not ask for the filter.
-func appendWaitingReplyClause(ctx context.Context, in ListActivitiesInput, arg func(any) int, where []string) ([]string, error) {
-	if in.WaitingReplyAsOf == nil {
-		return where, nil
-	}
-	clause, err := waitingReplyExistsClause(ctx, arg, *in.WaitingReplyAsOf, in.EntityType, in.EntityID)
-	if err != nil {
-		return nil, err
-	}
-	return append(where, clause), nil
-}
